+++ tests/conftest.py
@@ -17,6 +17,7 @@ Mocks server calls for testing
 """
 
 from base64 import b64encode
+from importlib.metadata import version
 import json
 import os
 import platform
@@ -24,7 +25,10 @@ import time
 from typing import Any, Optional
 from uuid import UUID
 
+from mockito import ANY, when
+from packaging.version import parse
 import pytest
+import requests
 from requests import HTTPError, Response
 
 from iqm.iqm_client import (
@@ -32,6 +36,9 @@ from iqm.iqm_client import (
     REQUESTS_TIMEOUT,
     Circuit,
     CircuitCompilationOptions,
+    DynamicQuantumArchitecture,
+    GateImplementationInfo,
+    GateInfo,
     HeraldingMode,
     Instruction,
     IQMClient,
@@ -70,6 +77,12 @@ def missing_run_id() -> UUID:
 
 @pytest.fixture()
 def sample_client(base_url) -> IQMClient:
+    client_version = parse(version('iqm-client'))
+    when(requests).get(f'{base_url}/info/client-libraries', headers=ANY, timeout=ANY).thenReturn(
+        MockJsonResponse(
+            200, {'iqm-client': {'min': f'{client_version.major}.0', 'max': f'{client_version.major + 1}.0'}}
+        )
+    )
     client = IQMClient(url=base_url)
     client._token_manager = None  # Do not use authentication
     return client
@@ -100,6 +113,11 @@ def existing_job_status_url(existing_job_url) -> str:
 @pytest.fixture()
 def quantum_architecture_url(base_url) -> str:
     return f'{base_url}/quantum-architecture'
+
+
+@pytest.fixture()
+def dynamic_architecture_url(base_url) -> str:
+    return f'{base_url}/api/v1/calibration/default/gates'
 
 
 @pytest.fixture
@@ -423,7 +441,7 @@ def deleted_status():
 
 
 @pytest.fixture
-def sample_quantum_architecture():
+def sample_static_architecture():
     return {
         'quantum_architecture': {
             'name': 'hercules',
@@ -440,79 +458,66 @@ def sample_quantum_architecture():
 
 
 @pytest.fixture
-def sample_move_architecture():
-    return {
-        'quantum_architecture': {
-            'name': 'hercules',
-            'qubits': ['COMP_R', 'QB1', 'QB2', 'QB3', 'COMP_R2'],
-            'qubit_connectivity': [
-                ['QB1', 'COMP_R'],
-                ['QB2', 'COMP_R'],
-                ['QB3', 'COMP_R'],
-                ['QB1', 'COMP_R2'],
-                ['QB2', 'COMP_R2'],
-                ['QB3', 'COMP_R2'],
-            ],
-            'operations': {
-                'prx': [['QB1'], ['QB2'], ['QB3']],
-                'cz': [['QB1', 'COMP_R'], ['QB2', 'COMP_R']],
-                'move': [['QB3', 'COMP_R']],
-                'measure': [['QB1'], ['QB2'], ['QB3']],
-                'barrier': [],
-            },
-        }
-    }
+def sample_dynamic_architecture():
+    return DynamicQuantumArchitecture(
+        calibration_set_id=UUID('26c5e70f-bea0-43af-bd37-6212ec7d04cb'),
+        qubits=['QB1', 'QB2', 'QB3'],
+        computational_resonators=[],
+        gates={
+            'prx': GateInfo(
+                implementations={
+                    'drag_gaussian': GateImplementationInfo(loci=(('QB1',), ('QB2',), ('QB3',))),
+                    'drag_crf': GateImplementationInfo(loci=(('QB1',), ('QB3',))),
+                },
+                default_implementation='drag_gaussian',
+                override_default_implementation={('QB3',): 'drag_crf'},
+            ),
+            'cz': GateInfo(
+                implementations={
+                    'tgss': GateImplementationInfo(loci=(('QB1', 'QB2'), ('QB1', 'QB3'))),
+                    'crf': GateImplementationInfo(loci=(('QB1', 'QB2'),)),
+                },
+                default_implementation='tgss',
+                override_default_implementation={},
+            ),
+            'measure': GateInfo(
+                implementations={'constant': GateImplementationInfo(loci=(('QB1',), ('QB2',)))},
+                default_implementation='constant',
+                override_default_implementation={},
+            ),
+        },
+    )
 
 
 @pytest.fixture
-def sample_dynamic_quantum_architecture():
-    return {
-        'calibration_set_id': 'cd4dd889-b88b-4370-ba01-eb8262ad9c53',
-        'qubits': ['QB1', 'QB2', 'QB3'],
-        'computational_resonators': ['COMP_R'],
-        'gates': {
-            'prx': {
-                'implementations': {
-                    'drag_gaussian': {
-                        'loci': [['QB1'], ['QB2'], ['QB3']],
-                    },
-                    'drag_crf': {
-                        'loci': [['QB1'], ['QB2'], ['QB3']],
-                    },
-                },
-                'default_implementation': 'drag_gaussian',
-                'override_default_implementation': {('QB3',): 'drag_crf'},
-            },
-            'cz': {
-                'implementations': {
-                    'tgss': {
-                        'loci': [['QB1', 'COMP_R'], ['QB3', 'COMP_R']],
-                    },
-                    'crf': {
-                        'loci': [['QB1', 'COMP_R']],
-                    },
-                },
-                'default_implementation': 'tgss',
-                'override_default_implementation': {},
-            },
-            'move': {
-                'implementations': {
-                    'tgss_crf': {
-                        'loci': [['QB2', 'COMP_R']],
-                    },
-                },
-                'default_implementation': 'tgss_crf',
-                'override_default_implementation': {},
-            },
-            'measure': {
-                'implementations': {
-                    'constant': {'loci': [['QB1'], ['QB2'], ['QB3']]},
-                },
-                'default_implementation': 'constant',
-                'override_default_implementation': {},
-            },
+def sample_move_architecture():
+    return DynamicQuantumArchitecture(
+        calibration_set_id=UUID('26c5e70f-bea0-43af-bd37-6212ec7d04cb'),
+        qubits=['QB1', 'QB2', 'QB3'],
+        computational_resonators=['COMP_R', 'COMP_R2'],
+        gates={
+            'prx': GateInfo(
+                implementations={'drag_gaussian': GateImplementationInfo(loci=(('QB1',), ('QB2',), ('QB3',)))},
+                default_implementation='drag_gaussian',
+                override_default_implementation={},
+            ),
+            'cz': GateInfo(
+                implementations={'tgss': GateImplementationInfo(loci=(('QB1', 'COMP_R'), ('QB2', 'COMP_R')))},
+                default_implementation='tgss',
+                override_default_implementation={},
+            ),
+            'move': GateInfo(
+                implementations={'tgss_crf': GateImplementationInfo(loci=(('QB3', 'COMP_R'),))},
+                default_implementation='tgss_crf',
+                override_default_implementation={},
+            ),
+            'measure': GateInfo(
+                implementations={'constant': GateImplementationInfo(loci=(('QB1',), ('QB2',), ('QB3',)))},
+                default_implementation='constant',
+                override_default_implementation={},
+            ),
         },
-    }
+    )
 
 
 class MockTextResponse:
@@ -539,6 +544,7 @@ class MockJsonResponse:
         self.status_code = status_code
         self.json_data = json_data
         self.history = history
+        self.url = 'https://example.com'
 
     @property
     def text(self):
@@ -568,18 +574,18 @@ def submit_failed_auth() -> MockJsonResponse:
 
 
 @pytest.fixture()
-def quantum_architecture_success(sample_quantum_architecture) -> MockJsonResponse:
-    return MockJsonResponse(200, sample_quantum_architecture)
+def static_architecture_success(sample_static_architecture) -> MockJsonResponse:
+    return MockJsonResponse(200, sample_static_architecture)
+
+
+@pytest.fixture()
+def dynamic_architecture_success(sample_dynamic_architecture) -> MockJsonResponse:
+    return MockJsonResponse(200, sample_dynamic_architecture.model_dump())
 
 
 @pytest.fixture()
 def move_architecture_success(sample_move_architecture) -> MockJsonResponse:
-    return MockJsonResponse(200, sample_move_architecture)
-
-
-@pytest.fixture()
-def dynamic_quantum_architecture_success(sample_dynamic_quantum_architecture) -> MockJsonResponse:
-    return MockJsonResponse(200, sample_dynamic_quantum_architecture)
+    return MockJsonResponse(200, sample_move_architecture.model_dump())
 
 
 @pytest.fixture()